rename DatastoreType -> Datastore, TransactorType -> Transactor
