+++ transactor/src/main/scala/io/mediachain/transactor/Copycat.scala
@@ -16,8 +16,8 @@ import cats.data.Xor
 object Copycat {
   import io.mediachain.transactor.StateMachine._
   import io.mediachain.types.Datastore
-  import io.mediachain.types.DatastoreTypes._
-  import io.mediachain.types.TransactorTypes._
+  import io.mediachain.types.Datastore._
+  import io.mediachain.types.Transactor._
 
 
   object Server {
