+++ transactor/src/main/scala/io/mediachain/transactor/Dummies.scala
@@ -4,7 +4,7 @@ import scala.collection.mutable.{HashMap => MHashMap, Map => MMap}
 
 object Dummies {
   import io.mediachain.types.Datastore
-  import io.mediachain.types.DatastoreTypes._
+  import io.mediachain.types.Datastore._
   import io.mediachain.util.cbor.CborAST._
 
   class DummyReference(val num: Int) extends Reference {
