+++ transactor/src/main/scala/io/mediachain/transactor/StateMachine.scala
@@ -12,9 +12,8 @@ import io.atomix.copycat.server.storage.snapshot.{SnapshotReader, SnapshotWriter
 import cats.data.Xor
 
 object StateMachine {
-  import io.mediachain.types.Datastore
-  import io.mediachain.types.DatastoreTypes._
-  import io.mediachain.types.TransactorTypes._
+  import io.mediachain.types.Datastore._
+  import io.mediachain.types.Transactor._
 
   val JournalBlockSize: Int = 4096 // blocksize for Journal Blockchain
