+++ transactor/src/test/scala/io/mediachain/transactor/JournalIndexSpec.scala
@@ -5,7 +5,7 @@ import java.util.concurrent.TimeUnit
 import scala.concurrent.Await
 import scala.concurrent.duration.Duration
 
-import io.mediachain.types.DatastoreTypes._
+import io.mediachain.types.Datastore._
 
 object JournalIndexSpec extends io.mediachain.BaseSpec
   with BeforeAll
